lipInBackEnd/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.121.2
git-filter-repo==2.47.0
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
openai==2.8.0
playwright==1.57.0
pydantic==2.12.4
pydantic_core==2.41.5
pyee==13.0.0
python-dotenv==1.2.1
python-multipart==0.0.9
PyPDF2==3.0.1
python-docx==1.1.2
Pillow==10.3.0
requests==2.32.5
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.49.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.38.0
